Override i18n_key on ActiveRecord-derived classes
This allows creating translations for the derived classes themselves
instead of polluting the base class translations.

For this we need to create an ActiveModel::Name, which uses the parent
class name, but derived class reference. We then override the i18n_key,
but not the other ones, to override i18n lookup but keep the same
parameter and route names.

A test is added to ensure the behavior

Fixes: #115