Fix a ruby warning due to redundant regex pattern
This fixes the following ruby warning:

 lib/active_type/virtual_attributes.rb:115:
 warning: character class has duplicated range: /\A[A-z0-9_]*\z/

This fix adapts the regex in a way that does not alter its matching pattern.
However, while the pattern `[A-z]` includes all the characters from the
alphabet, lower and upper case, it also includes the characters `[`, `\`, `]`,
and `^`. If this is not desired, the pattern should be changed to
`/\A[A-Za-z0-9_]*\z/` instead, as it would be more specific and does not allow
any additional characters.