Change regex pattern and fix ruby warning
This fixes the following ruby warning:

 lib/active_type/virtual_attributes.rb:115:
 warning: character class has duplicated range: /\A[A-z0-9_]*\z/

This fix adapts the regex in a way that alters its matching pattern. While the
pattern `[A-z]` included all the characters from the alphabet, lower and upper
case, it also included the characters `[`, `\`, `]`, `^` and `_`.

The new pattern makes this more specific, so that only alphanumeric characters
and `_` are permitted. This also fixes the ruby warning, as there is no more
redundancy.